Added ticket receiving class Changed the method for adding tickets Created a method for obtaining tickets Added to the cascade deletion structure

+++ DoubleDecker/src/main/java/de/doubledecker/doubledecker/repository/TicketRepository.java
@@ -1,8 +1,14 @@
 package de.doubledecker.doubledecker.repository;
 
+import de.doubledecker.doubledecker.controller.dto.TicketDTO;
 import de.doubledecker.doubledecker.domain.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
-public interface TicketRepository extends JpaRepository<Ticket,Integer> {
+import java.util.List;
+
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 
 }
+
+
